YH_Finance_API/code/requirements.txt
python-dotenv==1.0.1
requests==2.31.0
Flask-JWT-Extended==4.4.1
Flask==2.1.2
Flask-RESTful == 0.3.9
PyMySQL==1.0.3
Werkzeug==2.2.2